+++ src/main/java/redis/clients/jedis/BinaryShardedJedis.java
@@ -194,7 +194,7 @@ public class BinaryShardedJedis extends Sharded<Jedis, JedisShardInfo>
         Jedis j = getShard(key);
         return j.lpushx(key, string);
     }
-    
+
 
     public Long rpushx(byte[] key, byte[] string) {
         Jedis j = getShard(key);
@@ -206,27 +206,27 @@ public class BinaryShardedJedis extends Sharded<Jedis, JedisShardInfo>
 	return j.llen(key);
     }
 
-    public List<byte[]> lrange(byte[] key, int start, int end) {
+    public List<byte[]> lrange(byte[] key, long start, long end) {
 	Jedis j = getShard(key);
 	return j.lrange(key, start, end);
     }
 
-    public String ltrim(byte[] key, int start, int end) {
+    public String ltrim(byte[] key, long start, long end) {
 	Jedis j = getShard(key);
 	return j.ltrim(key, start, end);
     }
 
-    public byte[] lindex(byte[] key, int index) {
+    public byte[] lindex(byte[] key, long index) {
 	Jedis j = getShard(key);
 	return j.lindex(key, index);
     }
 
-    public String lset(byte[] key, int index, byte[] value) {
+    public String lset(byte[] key, long index, byte[] value) {
 	Jedis j = getShard(key);
 	return j.lset(key, index, value);
     }
 
-    public Long lrem(byte[] key, int count, byte[] value) {
+    public Long lrem(byte[] key, long count, byte[] value) {
 	Jedis j = getShard(key);
 	return j.lrem(key, count, value);
     }
@@ -286,7 +286,7 @@ public class BinaryShardedJedis extends Sharded<Jedis, JedisShardInfo>
 	return j.zadd(key, scoreMembers);
     }
 
-    public Set<byte[]> zrange(byte[] key, int start, int end) {
+    public Set<byte[]> zrange(byte[] key, long start, long end) {
 	Jedis j = getShard(key);
 	return j.zrange(key, start, end);
     }
@@ -311,17 +311,17 @@ public class BinaryShardedJedis extends Sharded<Jedis, JedisShardInfo>
 	return j.zrevrank(key, member);
     }
 
-    public Set<byte[]> zrevrange(byte[] key, int start, int end) {
+    public Set<byte[]> zrevrange(byte[] key, long start, long end) {
 	Jedis j = getShard(key);
 	return j.zrevrange(key, start, end);
     }
 
-    public Set<Tuple> zrangeWithScores(byte[] key, int start, int end) {
+    public Set<Tuple> zrangeWithScores(byte[] key, long start, long end) {
 	Jedis j = getShard(key);
 	return j.zrangeWithScores(key, start, end);
     }
 
-    public Set<Tuple> zrevrangeWithScores(byte[] key, int start, int end) {
+    public Set<Tuple> zrevrangeWithScores(byte[] key, long start, long end) {
 	Jedis j = getShard(key);
 	return j.zrevrangeWithScores(key, start, end);
     }
@@ -350,7 +350,7 @@ public class BinaryShardedJedis extends Sharded<Jedis, JedisShardInfo>
 	Jedis j = getShard(key);
 	return j.zcount(key, min, max);
     }
-    
+
     public Long zcount(byte[] key, byte[] min, byte[] max) {
 	Jedis j = getShard(key);
 	return j.zcount(key, min, max);
@@ -377,7 +377,7 @@ public class BinaryShardedJedis extends Sharded<Jedis, JedisShardInfo>
 	Jedis j = getShard(key);
 	return j.zrangeByScoreWithScores(key, min, max, offset, count);
     }
-    
+
 
     public Set<Tuple> zrangeByScoreWithScores(byte[] key, byte[] min, byte[] max) {
 	Jedis j = getShard(key);
@@ -412,7 +412,7 @@ public class BinaryShardedJedis extends Sharded<Jedis, JedisShardInfo>
 	Jedis j = getShard(key);
 	return j.zrevrangeByScoreWithScores(key, max, min, offset, count);
     }
-    
+
     public Set<byte[]> zrevrangeByScore(byte[] key, byte[] max, byte[] min) {
 	Jedis j = getShard(key);
 	return j.zrevrangeByScore(key, max, min);
@@ -436,7 +436,7 @@ public class BinaryShardedJedis extends Sharded<Jedis, JedisShardInfo>
 	return j.zrevrangeByScoreWithScores(key, max, min, offset, count);
     }
 
-    public Long zremrangeByRank(byte[] key, int start, int end) {
+    public Long zremrangeByRank(byte[] key, long start, long end) {
 	Jedis j = getShard(key);
 	return j.zremrangeByRank(key, start, end);
     }
