+++ src/main/java/redis/clients/jedis/Pipeline.java
@@ -10,22 +10,22 @@ import redis.clients.jedis.BinaryClient.LIST_POSITION;
 import redis.clients.jedis.exceptions.JedisDataException;
 
 public class Pipeline extends Queable {
-	
+
     private MultiResponseBuilder currentMulti;
-    
+
     private class MultiResponseBuilder extends Builder<List<Object>>{
     	private List<Response<?>> responses = new ArrayList<Response<?>>();
-		
+
 		@Override
 		public List<Object> build(Object data) {
 			@SuppressWarnings("unchecked")
 			List<Object> list = (List<Object>)data;
 			List<Object> values = new ArrayList<Object>();
-			
+
 			if(list.size() != responses.size()){
 				throw new JedisDataException("Expected data size " + responses.size() + " but was " + list.size());
 			}
-			
+
 			for(int i=0;i<list.size();i++){
 				Response<?> response = responses.get(i);
 				response.set(list.get(i));
@@ -43,7 +43,7 @@ public class Pipeline extends Queable {
     protected <T> Response<T> getResponse(Builder<T> builder) {
     	if(currentMulti != null){
     		super.getResponse(BuilderFactory.STRING); //Expected QUEUED
-    		
+
     		Response<T> lr = new Response<T>(builder);
     		currentMulti.addResponse(lr);
     		return lr;
@@ -52,9 +52,9 @@ public class Pipeline extends Queable {
     		return super.getResponse(builder);
     	}
     }
-	
+
     private Client client;
-    
+
     public void setClient(Client client) {
         this.client = client;
     }
@@ -76,14 +76,14 @@ public class Pipeline extends Queable {
      * pipeline. Whenever possible try to avoid using this version and use
      * Pipeline.sync() as it won't go through all the responses and generate the
      * right response type (usually it is a waste of time).
-     * 
+     *
      * @return A list of all the responses in the order you executed them.
      * @see sync
      */
     public List<Object> syncAndReturnAll() {
         List<Object> unformatted = client.getAll();
         List<Object> formatted = new ArrayList<Object>();
-        
+
         for (Object o : unformatted) {
             try {
             	formatted.add(generateResponse(o).get());
@@ -375,12 +375,12 @@ public class Pipeline extends Queable {
         return getResponse(BuilderFactory.STRING_SET);
     }
 
-    public Response<String> lindex(String key, int index) {
+    public Response<String> lindex(String key, long index) {
         client.lindex(key, index);
         return getResponse(BuilderFactory.STRING);
     }
 
-    public Response<String> lindex(byte[] key, int index) {
+    public Response<String> lindex(byte[] key, long index) {
         client.lindex(key, index);
         return getResponse(BuilderFactory.STRING);
     }
@@ -923,12 +923,12 @@ public class Pipeline extends Queable {
         return getResponse(BuilderFactory.LONG);
     }
 
-    public Response<Set<String>> zrange(String key, int start, int end) {
+    public Response<Set<String>> zrange(String key, long start, long end) {
         client.zrange(key, start, end);
         return getResponse(BuilderFactory.STRING_ZSET);
     }
 
-    public Response<Set<String>> zrange(byte[] key, int start, int end) {
+    public Response<Set<String>> zrange(byte[] key, long start, long end) {
         client.zrange(key, start, end);
         return getResponse(BuilderFactory.STRING_ZSET);
     }
@@ -943,7 +943,7 @@ public class Pipeline extends Queable {
             double max) {
         return zrangeByScore(key, toByteArray(min), toByteArray(max));
     }
-    
+
     public Response<Set<String>> zrangeByScore(String key, String min,
             String max) {
         client.zrangeByScore(key, min, max);
@@ -966,7 +966,7 @@ public class Pipeline extends Queable {
             double max, int offset, int count) {
         return zrangeByScore(key, toByteArray(min), toByteArray(max), offset, count);
     }
-    
+
     public Response<Set<String>> zrangeByScore(byte[] key, byte[] min,
     		byte[] max, int offset, int count) {
         client.zrangeByScore(key, min, max, offset, count);
@@ -983,7 +983,7 @@ public class Pipeline extends Queable {
             double max) {
         return zrangeByScoreWithScores(key, toByteArray(min), toByteArray(max));
     }
-    
+
     public Response<Set<Tuple>> zrangeByScoreWithScores(byte[] key, byte[] min,
     		byte[] max) {
         client.zrangeByScoreWithScores(key, min, max);
@@ -1001,7 +1001,7 @@ public class Pipeline extends Queable {
         client.zrangeByScoreWithScores(key, toByteArray(min), toByteArray(max), offset, count);
         return getResponse(BuilderFactory.TUPLE_ZSET);
     }
-    
+
     public Response<Set<Tuple>> zrangeByScoreWithScores(byte[] key, byte[] min,
     		byte[] max, int offset, int count) {
         client.zrangeByScoreWithScores(key, min, max, offset, count);
@@ -1043,7 +1043,7 @@ public class Pipeline extends Queable {
         client.zrevrangeByScore(key, toByteArray(max), toByteArray(min), offset, count);
         return getResponse(BuilderFactory.STRING_ZSET);
     }
-    
+
     public Response<Set<String>> zrevrangeByScore(byte[] key, byte[] max,
     		byte[] min, int offset, int count) {
         client.zrevrangeByScore(key, max, min, offset, count);
@@ -1061,7 +1061,7 @@ public class Pipeline extends Queable {
         client.zrevrangeByScoreWithScores(key, toByteArray(max), toByteArray(min));
         return getResponse(BuilderFactory.TUPLE_ZSET);
     }
-    
+
     public Response<Set<Tuple>> zrevrangeByScoreWithScores(byte[] key,
     		byte[] max, byte[] min) {
         client.zrevrangeByScoreWithScores(key, max, min);
@@ -1079,19 +1079,19 @@ public class Pipeline extends Queable {
         client.zrevrangeByScoreWithScores(key, toByteArray(max), toByteArray(min), offset, count);
         return getResponse(BuilderFactory.TUPLE_ZSET);
     }
-    
+
     public Response<Set<Tuple>> zrevrangeByScoreWithScores(byte[] key,
     		byte[] max, byte[] min, int offset, int count) {
         client.zrevrangeByScoreWithScores(key, max, min, offset, count);
         return getResponse(BuilderFactory.TUPLE_ZSET);
     }
 
-    public Response<Set<Tuple>> zrangeWithScores(String key, int start, int end) {
+    public Response<Set<Tuple>> zrangeWithScores(String key, long start, long end) {
         client.zrangeWithScores(key, start, end);
         return getResponse(BuilderFactory.TUPLE_ZSET);
     }
 
-    public Response<Set<Tuple>> zrangeWithScores(byte[] key, int start, int end) {
+    public Response<Set<Tuple>> zrangeWithScores(byte[] key, long start, long end) {
         client.zrangeWithScores(key, start, end);
         return getResponse(BuilderFactory.TUPLE_ZSET);
     }
@@ -1116,12 +1116,12 @@ public class Pipeline extends Queable {
         return getResponse(BuilderFactory.LONG);
     }
 
-    public Response<Long> zremrangeByRank(String key, int start, int end) {
+    public Response<Long> zremrangeByRank(String key, long start, long end) {
         client.zremrangeByRank(key, start, end);
         return getResponse(BuilderFactory.LONG);
     }
 
-    public Response<Long> zremrangeByRank(byte[] key, int start, int end) {
+    public Response<Long> zremrangeByRank(byte[] key, long start, long end) {
         client.zremrangeByRank(key, start, end);
         return getResponse(BuilderFactory.LONG);
     }
@@ -1135,30 +1135,30 @@ public class Pipeline extends Queable {
         client.zremrangeByScore(key, toByteArray(start), toByteArray(end));
         return getResponse(BuilderFactory.LONG);
     }
-    
+
     public Response<Long> zremrangeByScore(byte[] key, byte[] start, byte[] end) {
         client.zremrangeByScore(key, start, end);
         return getResponse(BuilderFactory.LONG);
     }
 
-    public Response<Set<String>> zrevrange(String key, int start, int end) {
+    public Response<Set<String>> zrevrange(String key, long start, long end) {
         client.zrevrange(key, start, end);
         return getResponse(BuilderFactory.STRING_ZSET);
     }
 
-    public Response<Set<String>> zrevrange(byte[] key, int start, int end) {
+    public Response<Set<String>> zrevrange(byte[] key, long start, long end) {
         client.zrevrange(key, start, end);
         return getResponse(BuilderFactory.STRING_ZSET);
     }
 
-    public Response<Set<Tuple>> zrevrangeWithScores(String key, int start,
-            int end) {
+    public Response<Set<Tuple>> zrevrangeWithScores(String key, long start,
+            long end) {
         client.zrevrangeWithScores(key, start, end);
         return getResponse(BuilderFactory.TUPLE_ZSET);
     }
 
-    public Response<Set<Tuple>> zrevrangeWithScores(byte[] key, int start,
-            int end) {
+    public Response<Set<Tuple>> zrevrangeWithScores(byte[] key, long start,
+            long end) {
         client.zrevrangeWithScores(key, start, end);
         return getResponse(BuilderFactory.TUPLE_ZSET);
     }
@@ -1284,39 +1284,39 @@ public class Pipeline extends Queable {
         client.flushDB();
         return getResponse(BuilderFactory.STRING);
     }
-    
+
     public Response<String> flushAll() {
         client.flushAll();
         return getResponse(BuilderFactory.STRING);
     }
-    
+
     public Response<String> info() {
         client.info();
         return getResponse(BuilderFactory.STRING);
     }
-    
+
     public Response<Long> dbSize() {
         client.dbSize();
         return getResponse(BuilderFactory.LONG);
     }
-    
+
     public Response<String> shutdown() {
         client.shutdown();
         return getResponse(BuilderFactory.STRING);
     }
-    
+
     public Response<String> ping() {
         client.ping();
         return getResponse(BuilderFactory.STRING);
     }
-    
+
     public Response<String> randomKey() {
         client.randomKey();
         return getResponse(BuilderFactory.STRING);
-    }   
-    
+    }
+
     public Response<String> select(int index){
     	client.select(index);
     	return getResponse(BuilderFactory.STRING);
-    }    
+    }
 }
